Don't unconditionally assume an offer exists in REMOVE_OFFER.
It is possible that the first offer we get from a remote node is from a
node that also has an Undefined level. In that case, the offer is "removed"
from the offer pool, but since we have never seen this node, the offer list
obviously doesn't contain the offer. Hence, we shouldn't assume that the
offer list always contains the given offer.